add detection of n
Former-commit-id: 018ec46003d5f619bea39e707b95615c27c0d677
Former-commit-id: e9a4acf495f16e063dbc9f3cb78f497bb02819c3
Former-commit-id: 5ff6c65e74f11e383fe7739c8144f0690cf31abc

+++ source/.chezmoiscripts/run_onchange_install-packages.sh
@@ -40,8 +40,10 @@ if [[ -z "$confirmation" || "${confirmation,,}" =~ ^\s*y(es)?\s*$ ]]; then
     xdg-utils zeal zip zoxide gnome-software gnome-software-plugin-flatpak
   # Maintain sudo after long install
   sudo -v
-  log 'installing nodejs. After getting node setup using `n`, run `sudo apt remove nodejs`'
-  install_if_available_apt nodejs
+  if ! is_accessible_cmd n; then
+    log 'installing nodejs. After getting node setup using `n`, run `sudo apt remove nodejs`'
+    install_if_available_apt nodejs
+  fi
   remove_if_installed_apt gnome-characters
 
   # Require custom ppas
